Move add-widgets form to a sub-folder
Since will need to add more related files
