+++ lib/assets/javascripts/cartodb3/editor.js
@@ -13,7 +13,7 @@ var VisDefinitionModel = require('./data/vis-definition-model');
 var WidgetDefinitionModel = require('./data/widget-definition-model');
 var ModalsServiceModel = require('./components/modals/modals-service-model');
 var viewFactory = require('./components/view-factory');
-var AddWidgetsView = require('./editor/add-widgets-view');
+var AddWidgetsView = require('./editor/add-widgets/add-widgets-view');
 
 // JSON data passed from entry point (editor/visualizations/show.html):
 var vizJSON = window.vizJSON;
