+++ lib/assets/javascripts/cartodb3/editor/add-widgets/add-widgets-view.js
