+++ lib/assets/test/spec/cartodb3/editor/add-widgets/add-widgets-view.spec.js
@@ -1,6 +1,6 @@
-var AddWidgetsView = require('../../../../javascripts/cartodb3/editor/add-widgets-view');
+var AddWidgetsView = require('../../../../../javascripts/cartodb3/editor/add-widgets/add-widgets-view');
 
-describe('editor/add-widgets-view', function () {
+describe('editor/add-widgets/add-widgets-view', function () {
   beforeEach(function () {
     this.modalmodel = {};
     this.view = new AddWidgetsView({
